# Conflicts:
#	API/routes/event_routes.py
#	SPLMobile/androidApp/src/main/java/com/example/splmobile/android/ui/main/screens/events/CreateEventScreen.kt

+++ SPLMobile/androidApp/src/main/java/com/example/splmobile/android/ui/main/screens/garbageSpots/GarbageSpotInfoScreen.kt
@@ -39,7 +39,7 @@ fun GarbageSpotInfoScreen(
     LaunchedEffect(Unit) {
 
         //garbageSpotViewModel.getGarbageSpots(authViewModel.tokenState.value)
-        garbageSpotViewModel.getGarbageSpotById(garbageSpotId!!,authViewModel.tokenState.value)
+        garbageSpotViewModel.getGarbageSpotById(garbageSpotId!!.toLong(),authViewModel.tokenState.value)
 
 
     }
